app/requirements.txt
streamlit
pyats
langgraph
langchain
langchain-openai
python-dotenv
requests